requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.1.0
autoflake==2.2.1
black==23.12.1
cfgv==3.4.0
click==8.1.7
colorlog==6.8.0
distlib==0.3.8
fastapi==0.108.0
filelock==3.13.1
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.15.0
hyperframe==6.0.1
identify==2.5.33
idna==3.6
isort==5.13.2
multidict==6.0.4
mypy==1.8.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pre-commit==3.6.0
priority==2.0.0
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.1.0
PyMySQL==1.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.24
SQLAlchemy-serializer==1.4.1
starlette==0.32.0.post1
typing_extensions==4.9.0
virtualenv==20.25.0
wsproto==1.2.0
yarl==1.9.4
